Refactor Command class into subclasses and reorganize packages
For improving code maintainability and separation of concerns

* Split to subclasses to handle different command types.
* Combined subclasses and Command class into command package.
* Moved other classes into a new utility package.
