+++ src/main/java/echolex/utility/Parser.java
@@ -1,4 +1,4 @@
-package echolex.command;
+package echolex.utility;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -10,6 +10,12 @@ import java.time.temporal.ChronoField;
 import java.util.Arrays;
 import java.util.HashMap;
 
+import echolex.command.AddCommand;
+import echolex.command.Command;
+import echolex.command.DeleteCommand;
+import echolex.command.FindCommand;
+import echolex.command.ListCommand;
+import echolex.command.MarkCommand;
 import echolex.error.EchoLexException;
 
 /**
@@ -44,7 +50,37 @@ public class Parser {
             options.put(optionParts[0], optionParts[1]);
         }
 
-        return new Command(command, argument, options);
+        return identifyCommandType(command, argument, options);
+
+    }
+
+    /**
+     * Identifies the command type.
+     *
+     * @param command Command string.
+     * @param argument Argument string.
+     * @param options Options hashmap.
+     * @return Command object.
+     */
+    public static Command identifyCommandType(String command, String argument, HashMap<String, String> options) {
+
+        switch (command) {
+        case "list":
+            return new ListCommand(command, argument, options);
+        case "find":
+            return new FindCommand(command, argument, options);
+        case "mark":
+        case "unmark":
+            return new MarkCommand(command, argument, options);
+        case "delete":
+            return new DeleteCommand(command, argument, options);
+        case "todo":
+        case "deadline":
+        case "event":
+            return new AddCommand(command, argument, options);
+        default:
+            return new Command(command, argument, options);
+        }
 
     }
 
@@ -72,7 +108,8 @@ public class Parser {
             return date.atStartOfDay(); // may be modified to accommodate time in the future
 
         } catch (DateTimeParseException e) {
-            throw new EchoLexException("Invalid date: " + e.getMessage() + "\nPlease provide the date in yyyy-mm-dd format (e.g., 2019-10-15)");
+            throw new EchoLexException("Invalid date: " + e.getMessage()
+                    + "\nPlease provide the date in yyyy-mm-dd format (e.g., 2019-10-15)");
         }
 
     }
