+++ src/main/java/echolex/utility/Storage.java
@@ -1,4 +1,4 @@
-package echolex.command;
+package echolex.utility;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
